+++ FuzzyAutocomplete/DVTTextCompletionListWindowController+FuzzyAutocomplete.m
@@ -145,7 +145,9 @@ const char kRowHeightKey;
 // We modify titleColumnX with score column's width, so that title column is aligned with typed text
 - (CGRect) _fa_preferredWindowFrameForTextFrame: (CGRect) textFrame columnsWidth: (double *) widths titleColumnX: (double) titleX {
     NSTableView * tableView = [self valueForKey: @"_completionsTableView"];
-    return [self _fa_preferredWindowFrameForTextFrame: textFrame columnsWidth: widths titleColumnX: titleX + [self _fa_widthForScoreColumn] + tableView.intercellSpacing.width];
+    NSTableColumn * scoreColumn = [tableView tableColumnWithIdentifier: @"score"];
+    CGFloat additionalWidth = scoreColumn ? [self _fa_widthForScoreColumn] + tableView.intercellSpacing.width : 0;
+    return [self _fa_preferredWindowFrameForTextFrame: textFrame columnsWidth: widths titleColumnX: titleX + additionalWidth];
 }
 
 
@@ -226,10 +228,10 @@ const char kRowHeightKey;
 - (CGFloat) _fa_widthForScoreColumn {
     NSTableView * tableView = [self valueForKey: @"_completionsTableView"];
     NSTableColumn * scoreColumn = [tableView tableColumnWithIdentifier: @"score"];
-    if (scoreColumn && self.session.fa_nonZeroScores) {
+    DVTFontAndColorTheme * theme = [DVTFontAndColorTheme currentTheme];
+    if (scoreColumn && self.session.fa_nonZeroScores && theme.sourcePlainTextFont) {
         NSNumberFormatter * formatter = ((NSCell *)scoreColumn.dataCell).formatter;
         NSString * sampleValue = [formatter stringFromNumber: @0];
-        DVTFontAndColorTheme * theme = [DVTFontAndColorTheme currentTheme];
         NSDictionary * attributes = @{ NSFontAttributeName : theme.sourcePlainTextFont };
         return [[NSAttributedString alloc] initWithString: sampleValue attributes: attributes].size.width + 6;
     } else {